+++ interface/src/lib/components/admin/settings/common/AssetSelect.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { createEventDispatcher } from "svelte";
   import { _ } from "svelte-i18n";
-  import toast from "svelte-french-toast";
+import { toast } from "svelte-sonner";
   import { mixinAsset } from "$lib/helpers/mixin/mixin";
   import { MIXIN_API_BASE_URL } from "$lib/helpers/constants";